PTCD-636 Archive instead of delete for Functions/DeleteCoursesByUKPRN
We *think* that hard delete is causing changes to not show up in the
CosmosDB Change Feed Listener (CFL), and as such result in data
inconsistency and errors across the system.

There are no tests in the patch for this change in behaviour because the
existing system architecture has made this virtually impossible without
major changes. This has been discussed at length with the team and we
are resigned to moving forward without test coverage in this instance.

I'm informed by fellow devs that course status is calculated based on
aggregation of course run statuses, so shouldn't be any need to set the
status on the actual course. Ideally we'd wnat to clean up the models to
prevent inconsistent states, but given the state of the project that
will be a big task.