PTCD-636 Archive instead of delete for Functions/DeleteBulkUploadCourses
We *think* that hard delete is causing changes to not show up in the
CosmosDB Change Feed Listener (CFL), and as such result in data
inconsistency and errors across the system.

There are no tests in the patch for this change in behaviour because the
existing system architecture has made this virtually impossible without
major changes. This has been discussed at length with the team and we
are resigned to moving forward without test coverage in this instance.

* Copy-paste of `ArchiveCourseRunsByUKPRN` to
  `ArchivePendingBulkUploadCourseRunsByUKPRN` because we're making this
  change blind and this is most likely to work.
* `course.CourseStatus` filtering deleted code in CosmosDbHelper